# GH-3996 Reduce locking in AbstractSailConnection