@hmottestad 2022-06-19
## Before
```
```

## After
```
Benchmark                              Mode  Cnt  Score   Error  Units
ConcurrentQueryBenchmark.hasStatement  avgt    5  8.025 ± 0.237  ms/op
```