@hmottestad 2022-06-26
## Before
```
Benchmark                                               Mode  Cnt  Score   Error  Units
ConcurrentQueryBenchmark.getNamespaces                  avgt    5  0.935 ± 0.095  ms/op
ConcurrentQueryBenchmark.getNamespacesSharedConnection  avgt    5  4.269 ± 0.097  ms/op
ConcurrentQueryBenchmark.hasStatement                   avgt    5  0.990 ± 0.045  ms/op
ConcurrentQueryBenchmark.hasStatementSharedConnection   avgt    5  2.988 ± 0.233  ms/op
```

## After
```
Benchmark                                               Mode  Cnt  Score   Error  Units
ConcurrentQueryBenchmark.getNamespaces                  avgt    5  0.908 ± 0.042  ms/op
ConcurrentQueryBenchmark.getNamespacesSharedConnection  avgt    5  1.473 ± 0.057  ms/op
ConcurrentQueryBenchmark.hasStatement                   avgt    5  0.914 ± 0.115  ms/op
ConcurrentQueryBenchmark.hasStatementSharedConnection   avgt    5  1.010 ± 0.052  ms/op
```